+++ tests/unit/hg_oap/dates/test_calendar.py
@@ -2,6 +2,7 @@ from datetime import date
 
 import pytest
 
+from hg_oap.dates import CalendarImpl, DelegateCalendar, UnionCalendar
 from hg_oap.dates.calendar import WeekendCalendar, HolidayCalendar
 from hg_oap.dates.dgen import weeks
 
@@ -19,6 +20,7 @@ from hg_oap.dates.dgen import weeks
 def test_weekend_calendar_add_business_days(d, t, r):
     assert WeekendCalendar([5, 6]).add_business_days(d, t) == r
 
+
 @pytest.mark.parametrize(['d', 't', 'r'], (
         (date(1998, 2, 3), 1, date(1998, 2, 2)),
         (date(1998, 2, 3), 4, date(1998, 1, 28)),
@@ -31,6 +33,7 @@ def test_weekend_calendar_add_business_days(d, t, r):
 ))
 def test_weekend_calendar_sub_business_days(d, t, r):
     assert WeekendCalendar([5, 6]).sub_business_days(d, t) == r
+
 
 @pytest.mark.parametrize(['d', 't', 'r'], (
         (date(1998, 2, 3), 1, date(1998, 2, 4)),
@@ -47,6 +50,7 @@ def test_holiday_calendar_add_business_days(d, t, r):
     calendar = HolidayCalendar(tuple(all_fridays()))
     assert calendar.add_business_days(d, t) == r
 
+
 @pytest.mark.parametrize(['d', 't', 'r'], (
         (date(1998, 2, 3), 1, date(1998, 2, 4)),
         (date(1998, 2, 3), 4, date(1998, 2, 10)),
@@ -61,6 +65,7 @@ def test_holiday_calendar_add_business_days_1(d, t, r):
     all_mondays = '1997-01-01' <= weeks.mon <= '1999-01-01'
     calendar = HolidayCalendar(tuple(all_mondays()))
     assert calendar.add_business_days(d, t) == r
+
 
 @pytest.mark.parametrize(['d', 't', 'r'], (
         (date(1998, 2, 3), 1, date(1998, 2, 2)),
@@ -98,3 +103,24 @@ def test_holiday_calendar_is_holiday():
     calendar = HolidayCalendar((date(2024, 3, 29), date(2024, 4, 1)))
     assert calendar.is_holiday(date(2024, 3, 29))
     assert not calendar.is_holiday(date(2024, 3, 28))
+
+
+def test_calendar_impl():
+    cal = CalendarImpl([date(2024, 3, 29), date(2024, 4, 1)])
+    assert cal.is_business_day(date(2024, 3, 28))
+    assert not cal.is_business_day(date(2024, 3, 29))
+
+
+def test_calendar_delegate():
+    cal = DelegateCalendar(CalendarImpl([date(2024, 3, 29), date(2024, 4, 1)]))
+    assert cal.is_business_day(date(2024, 3, 28))
+    assert not cal.is_business_day(date(2024, 3, 29))
+
+
+def test_union_calendar():
+    cal = UnionCalendar(CalendarImpl([date(2024, 3, 29)]),
+                        CalendarImpl([date(2024, 4, 1)]))
+
+    assert cal.is_business_day(date(2024, 3, 28))
+    assert not cal.is_business_day(date(2024, 4, 1))
+    assert not cal.is_business_day(date(2024, 3, 29))